design/ux feedback: * otel trace span color * move otel attrs to step into panel
